% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_X.R, R/Simulate_Y.R
\name{from_south}
\alias{from_south}
\title{Displacement of the vehicles from south}
\usage{
from_south(i, x, y)

from_south(i, x, y)
}
\arguments{
\item{i, x, y}{i is the timing of the overall simulation, x,y correspond
to the current position of the vehicle.}
}
\value{
No return value

No return value
}
\description{
This procedure is used inside the procedures;
move_bicycle()
move_bike()
move_car()
move_truck()
it indicate the vehicle coming from the south and it shows it current position.
this procedure is not accessible by the public
this procedure is automatically call once the simulation of a particular vehicle is lunched.

This procedure is used inside the procedures;
move_bicycle()
move_bike()
move_car()
move_truck()
it indicate the vehicle coming from the south and it shows it current position.
this procedure is not accessible by the public
this procedure is automatically call once the simulation of a particular vehicle is lunched.
}
