% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_X.R, R/Simulate_Y.R
\name{light_color_red_left}
\alias{light_color_red_left}
\title{Creation of the traffic light}
\usage{
light_color_red_left(x, y)

light_color_red_left(x, y)
}
\arguments{
\item{are}{the coordinates of the traffic light on the roads in other word its location.}
}
\value{
No return value

No return value
}
\description{
This procedure creates the traffic light  on the 4 way direction road and it is incharge of the
red light.
 this procedure is use in the other procedure and is not accessible by the public
this procedure is automatically call once the simulation of a particular vehicle is lunched.

This procedure creates the traffic light  on the 4 way direction road and it is incharge of the
red light.
 this procedure is use in the other procedure and is not accessible by the public
this procedure is automatically call once the simulation of a particular vehicle is lunched.
}
