% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_X.R, R/Simulate_Y.R
\name{move_bicycle}
\alias{move_bicycle}
\title{Displacement of a bicycle}
\usage{
move_bicycle(simulation_time, direction_from, direction_to)

move_bicycle(simulation_time, direction_from, direction_to)
}
\arguments{
\item{simulation_time, direction_from, direction_to}{simulation_time is the duration of the simulation. The simulation time must be an integer ranging from 100 upward.
For a better simulation, use 200 for the timing.
direction_from is the initial position of the vehicle, the acceptable input are;
"EAST", "NORTH", "SOUTH"
direction_to is the final destination of the vehicle or the direction the vehicle will
take once it reaches the junction. The acceptable input are; "EAST", "NORTH", "SOUTH"}
}
\value{
No return value

No return value
}
\description{
This procedure shows the simulation of a bicycle coming from any of the four direction of the road
move_east()
move_west()
move_north()
move_south()

This procedure shows the simulation of a bicycle coming from any of the four direction of the road
move_east()
move_north()
move_south()
}
